requirements.txt
pillow>=6.0.0